updated user handling

+++ db/migrate/20120324062537_create_attendees.rb
@@ -2,7 +2,7 @@ class CreateAttendees < ActiveRecord::Migration
   def change
     create_table :attendees do |t|
       t.string :name
-      t.integer :hourly_rate
+      t.decimal :hourly_rate
       t.string :email
       t.belongs_to :user
       t.references :meeting
